Feature: Dashboard search
Related to #20

There is an API endpoint that accepts a query string to search available dashboards. Rather than re-invent the wheel, this adds two new calls to leverage that endpoint. This prevents a user from haivng to always use `hubot graf list` to get the entire list of the dashboards in order to query one of them.

- `hubot graf search <term>` will translate to `/api/search/?query=<term>` and return dashboards that have titles that match specified term
- `hubot graf list <tag>` will translate to `/api/search/?tag=<tag>` and return dashboards that match the specified tag