Add CONTRIBUTING.md, update README on config vars
As mentioned in #35, the configuration options for S3 were getting a bit out of hand. This PR attempts to clarify things a bit and make it easier for a new or upgrading user to understand what all is required to get the package up and running.

Also, I added a CONTRIBUTING.md file with some notes on working with the package. My rationale there is to make it less of a hurdle to get started hacking on this project.