Grafana v5.0 support
Fixes #66

This refactors the process by which the Hubot package locates the
relevant data in the API response. The original issue came from the
elimination of the `rows` property, replacing it with a coordinate
system instead.

Note that this will also drop support for versions < 2.0.2 to keep the
code a bit cleaner. Users should pin to `v1.8.0` if you are using an
older version of Grafana.